+++ mynlp/src/main/java/com/mayabot/nlp/perceptron/PerceptronFileFormat.kt
@@ -46,7 +46,6 @@ object PerceptronFileFormat {
     }
 
     @JvmStatic
-    @JvmOverloads
     fun loadFromNlpResource(prefix: String, nlpEnv: MynlpEnv): PerceptronModel {
 
         val parameter = nlpEnv.loadResource("$prefix/parameter.bin")