+++ mynlp/src/test/java/com/mayabot/nlp/Highlight.kt
@@ -1,0 +1,37 @@
+package com.mayabot.nlp
+
+import com.mayabot.nlp.module.Highlighter
+import com.mayabot.nlp.module.highlight
+import org.junit.Assert
+import org.junit.Test
+
+class HighlightTest {
+
+    val words = listOf<String>("居住证", "居住", "住宅")
+
+    @Test
+    fun test() {
+
+        val highlighter = Highlighter(words)
+        val text = "这个居住证，怎么办，居住和住宅----"
+
+        Assert.assertEquals(highlighter.replace(text), "这个<em>居住证</em>，怎么办，<em>居住</em>和<em>住宅</em>----")
+    }
+
+    @Test
+    fun test2() {
+        val highlighter = Highlighter(words, "div")
+        val text = "这个居住证，怎么办，居住和住宅----"
+
+        Assert.assertEquals(highlighter.replace(text), "这个<div>居住证</div>，怎么办，<div>居住</div>和<div>住宅</div>----")
+    }
+
+    @Test
+    fun test3() {
+        val text = "这个居住证，怎么办，居住和住宅----"
+
+        val result = text.highlight(words)
+
+        Assert.assertEquals(result, "这个<em>居住证</em>，怎么办，<em>居住</em>和<em>住宅</em>----")
+    }
+}